# api/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/admin/sync/posts": {
            "post": {
                "description": "将已发布文章写入 Hexo source/_posts，可选执行 hexo 构建命令",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Hexo 文章全量同步",
                "parameters": [
                    {
                        "description": "选项",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.SyncPostsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.SyncResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/sync/status": {
            "get": {
                "description": "查询上次同步时间与本地 beehive 文章文件数量等",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Hexo 同步状态",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.SyncStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.SyncPostsRequest": {
            "type": "object",
            "properties": {
                "rebuild": {
                    "description": "Rebuild 为 true 且服务端配置了 hexo.generate_args 时，在同步后执行静态站点生成。",
                    "type": "boolean"
                }
            }
        },
        "github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.SyncResponse": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "deleted": {
                    "type": "integer"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total": {
                    "type": "integer"
                },
                "updated": {
                    "type": "integer"
                }
            }
        },
        "github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.SyncStatusResponse": {
            "type": "object",
            "properties": {
                "last_sync_time": {
                    "type": "string"
                },
                "local_files": {
                    "type": "integer"
                },
                "pending_sync": {
                    "type": "boolean"
                },
                "total_posts": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}